requirements.txt
# Basic dependencies - Optimized for Python 3.11 compatibility
numpy>=1.21.0,<1.25.0
pandas>=1.4.0,<2.0.0
scipy>=1.8.0,<1.11.0
setuptools>=60.0.0
Cython>=0.29.0,<3.0.0
joblib>=1.1.0
scikit-learn>=1.0.0,<1.3.0
statsmodels>=0.13.0,<0.15.0

# Main packages
alpha-vantage>=2.3.1
yfinance>=0.2.0
requests>=2.28.0
pmdarima>=2.0.4
matplotlib>=3.6.0
python-dotenv>=1.0.0

# Backtesting Framework
backtesting>=0.3.3
